# Improve `select` row perf for large N