# Description
While reviewing #10350 I noticed that a `HashSet<usize>` was used to deduplicate the incoming rows, which are then sorted after cloning to a separate `Vec`. This sounds like a candidate for a `BTreeSet` which guarantees the ordering.

In the process I removed some dead code.

- Use `BTreeSet` instead of `HashSet`
- Remove dead `skip` logic
- Use `BTreeSet` directly in `NthIterator`
# User-Facing Changes
None

# Benchmarking

Data was cached and test script run for `cargo build --release`

Random data was chosen to create some duplication and create a worst case amount of work.
No significant impact was observed for small data with predictable input.

```
use std repeat

# let data = ls | repeat 1000 | flatten
# let index = random dice -d 10000 -s 20000
let data = open data.nuon
let index = open index.nuon

1..1000 | each { timeit { $data | select $index } } | math avg
```

```
=== bbea7da66 ===
17ms 96µs 202ns
=== 8b5d50968 ===
17ms 906µs 870ns
=== 94d651f21 ===
17ms 70µs 277ns
=== c085d27fa ===
14ms 429µs 701ns
=== ddcb22e22 ===
14ms 349µs 939ns
```