@Tiggax 2023-09-13
`reject` uses a similar operation if i remember correctly. 
Would maybe refactoring it after this PR be benefitial?
maybe I'll do it tommorow :thinking: