@sholderbach 2023-09-21
# Benchmarking results

## Small case
`select 42` for a table of 100 lines
```
=== bbea7da66 ===
25µs 787ns
=== 8b5d50968 ===
25µs 453ns
=== 94d651f21 ===
26µs 172ns
=== c085d27fa ===
25µs 329ns
=== ddcb22e22 ===
25µs 49ns
=== fad26c1de ===
26µs 952ns
```

## Big case
see setup above
```
=== bbea7da66 ===
15ms 569µs 270ns
=== 8b5d50968 ===
15ms 995µs 989ns
=== 94d651f21 ===
16ms 232µs 336ns
=== c085d27fa ===
13ms 573µs 916ns
=== ddcb22e22 ===
13ms 293µs 353ns
=== fad26c1de ===
13ms 133µs 17ns
```